+++ lib/finance_chain/supervisor.ex
@@ -1,0 +1,16 @@
+defmodule FinanceChain.BlockChain.Supervisor do
+  use Supervisor
+
+  def start_link(opts) do
+    Supervisor.start_link(__MODULE__, :ok, opts)
+  end
+
+  @impl true
+  def init(:ok) do
+    children = [
+      {FinanceChain.BlockChain.Server,name: FinanceChain.BlockChain.Server}
+    ]
+
+    Supervisor.init(children, strategy: :one_for_one)
+  end
+end
